+++ CHANGELOG.md
@@ -6,13 +6,14 @@ Services and Managers
 
 - `KeychainService` API is changed from `Result` types to throwing methods
 - `KeychainServiceError` is now an `enum`
+- `SessionManager` is now an `actor`
 
-Helpers
+Models
 
+- `ObservableContainer` is added
 - `AtomicContainer` is added
 - `AtomicInteger` is now an `actor`
 - `AtomicInteger` now contains many methods for incrementation and decrementation
-- `SessionManager` is now an `actor`
 
 Architectural Pattern Helpers